requirements.txt
altgraph==0.17.3
asttokens==2.0.8
async-generator==1.10
attrs==22.1.0
backcall==0.2.0
black==22.8.0
bottle==0.12.25
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
decorator==5.1.1
defusedxml==0.7.1
et-xmlfile==1.1.0
exceptiongroup==1.0.0rc9
executing==1.1.1
future==0.18.2
h11==0.14.0
idna==3.4
ipdb==0.13.9
ipython==8.5.0
jedi==0.18.1
matplotlib-inline==0.1.6
mypy-extensions==0.4.3
numpy==1.23.3
odfpy==1.4.1
openpyxl==3.0.10
outcome==1.2.0
packaging==22.0
pandas==1.5.0
parso==0.8.3
pathspec==0.10.1
pefile==2022.5.30
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.5.2
prompt-toolkit==3.0.31
proxy-tools==0.1.0
ptyprocess==0.7.0
pure-eval==0.2.2
pycairo==1.23.0
pycparser==2.21
Pygments==2.13.0
PyGObject==3.44.1
pyinstaller==5.8.0
pyinstaller-hooks-contrib==2022.10
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
PyQtWebEngine==5.15.6
PyQtWebEngine-Qt5==5.15.2
PySide2==5.15.2.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.4
pywebview==4.0.2
pywin32-ctypes==0.2.0
QtPy==2.3.1
requests==2.28.1
selenium==4.7.2
shiboken2==5.15.2.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
stack-data==0.5.1
toml==0.10.2
tomli==2.0.1
tqdm==4.64.1
traitlets==5.4.0
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.12
wcwidth==0.2.5
webdriver-manager==3.8.5
wsproto==1.2.0
